backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
asyncpg>=0.29.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.7.0

# Dokumentum feldolgozás
pymupdf>=1.24.0
pdfminer.six>=20221105
python-docx>=1.1.0
openpyxl>=3.1.0

# Embedding-alapú hasonlóság (learning service)
numpy>=1.26.0
